lab 1 and 2

+++ src/LAB2/task1/Calculator.java
@@ -37,12 +37,12 @@ public class Calculator {
         return a * b;
     }
 
-    public int divide(int a, int b) {
-        return a / b;
+    public double divide(int a, int b) {
+        return (double)a / b;
     }
 
-    public long divide(long a, long b) {
-        return a / b;
+    public double divide(long a, long b) {
+        return (double)a / b;
     }
 
     public double divide(double a, double b) {
